wip: move bottom nav review logic to list-mobile.vue
