+++ apps/survey/src/components/layouts/meal-list/list-mobile.vue
@@ -1,26 +1,33 @@
 <template>
-  <v-card class="meal-list-mobile" flat tile>
-    <v-toolbar bottom class="sticky_toolbar" flat>
-      <v-tabs center-active height="56px" icons-and-text slider-size="4" touch :value="activeTab">
-        <v-tabs-slider color="secondary"></v-tabs-slider>
-        <v-tab v-for="meal in meals" :key="meal.id" @click="mealSelected(meal.id)">
-          <v-badge
-            bordered
-            color="grey"
-            :content="meal.foods.length"
-            left
-            :value="!!meal.foods.length"
-          >
-            <p v-if="meal.time">{{ getMealTime(meal) }}</p>
-            <p v-else>
-              <v-icon x-small>$question</v-icon>
-            </p>
-            {{ translate(meal.name) }}
-          </v-badge>
-        </v-tab>
-      </v-tabs>
-    </v-toolbar>
-  </v-card>
+  <v-bottom-sheet>
+    <template #activator="{ on, attrs }">
+      <v-btn value="review" v-bind="attrs" v-on="on">
+        <span class="text-overline font-weight-medium">
+          {{ $t('recall.actions.nav.review') }}
+        </span>
+        <v-icon class="pb-1">$survey</v-icon>
+      </v-btn>
+      <v-divider vertical></v-divider>
+    </template>
+    <v-list>
+      <v-subheader>{{ $t('recall.menu.title') }}</v-subheader>
+      <v-list class="meal-list pt-0" dense flat tile>
+        <template v-for="meal in meals">
+          <component
+            :is="expandable ? 'meal-item-expandable' : 'meal-item'"
+            :key="meal.id"
+            :meal="meal"
+            :selected="selectedMealId === meal.id"
+            :selected-food-id="selectedFoodId"
+            :selected-food-in-meal="isSelectedFoodInMeal(meal.id)"
+            @action="action"
+            @food-selected="foodSelected"
+            @meal-selected="mealSelected"
+          ></component>
+        </template>
+      </v-list>
+    </v-list>
+  </v-bottom-sheet>
 </template>
 
 <script lang="ts">
@@ -33,12 +40,21 @@ import type { MealState } from '@intake24/common/types';
 import { useI18n } from '@intake24/i18n';
 import { useMealUtils } from '@intake24/survey/composables';
 import { useSurvey } from '@intake24/survey/stores';
-import { getMealIndex } from '@intake24/survey/util';
+import { getFoodIndexRequired } from '@intake24/survey/util';
+
+import MealItem from './meal-item.vue';
+import MealItemExpandable from './meal-item-expandable.vue';
 
 export default defineComponent({
   name: 'MealListMobile',
 
+  components: { MealItem, MealItemExpandable },
+
   props: {
+    expandable: {
+      type: Boolean,
+      default: false,
+    },
     meals: {
       type: Array as PropType<MealState[]>,
       required: true,
@@ -55,26 +71,69 @@ export default defineComponent({
   },
 
   computed: {
-    ...mapState(useSurvey, ['selectedMealOptional']),
+    ...mapState(useSurvey, ['selection']),
 
-    activeTab() {
-      if (this.selectedMealOptional === undefined) return 0;
-      return getMealIndex(this.meals, this.selectedMealOptional.id) ?? 0;
+    selectedMealId() {
+      if (this.selection.element?.type !== 'meal') return undefined;
+      return this.selection.element.mealId;
+    },
+    selectedFoodId() {
+      if (this.selection.element?.type !== 'food') return undefined;
+      return this.selection.element.foodId;
     },
   },
 
   methods: {
-    mealSelected(mealId: string) {
-      this.action('selectMeal', mealId);
-    },
+    isSelectedFoodInMeal(mealId: string): boolean {
+      if (this.selection.element?.type !== 'food') return false;
 
+      const foodIndex = getFoodIndexRequired(this.meals, this.selection.element.foodId);
+
+      return this.meals[foodIndex.mealIndex].id === mealId;
+    },
     action(type: FoodActionType | MealActionType, id?: string) {
       this.$emit('action', type, id);
+    },
+    foodSelected(foodId: string) {
+      this.action('selectFood', foodId);
+    },
+    mealSelected(mealId: string) {
+      this.action('selectMeal', mealId);
     },
   },
 });
 </script>
 
-<style lang="scss" scoped>
-@import '../../../scss/meallistmobile.scss';
+<style lang="scss">
+@import 'vuetify/src/styles/styles.sass';
+
+.meal-list {
+  .selected {
+    background: map-get($orange, 'lighten-5');
+  }
+
+  .selected-food {
+    background: map-get($orange, 'lighten-4');
+  }
+
+  .v-list-group__header {
+    padding-left: 12px !important;
+  }
+
+  .v-list-item:hover {
+    background: map-get($orange, 'lighten-5');
+  }
+
+  .v-list-group--active > .v-list-group__header > .v-list-group__header__prepend-icon {
+    margin-right: 6px;
+
+    .v-icon {
+      transform: rotate(180deg);
+    }
+  }
+
+  .v-list-group__header__append-icon {
+    display: none !important;
+  }
+}
 </style>
