+++ apps/survey/src/components/layouts/meal-list/meal-item.vue
@@ -19,12 +19,14 @@
         </v-tooltip>
       </v-list-item-action>
       <v-list-item-action class="my-auto">
-        <context-menu
-          :entity="meal"
-          :entity-name="mealName"
-          v-bind="{ menu }"
-          @action="action"
-        ></context-menu>
+        <v-expand-transition>
+          <context-menu
+            :entity="meal"
+            :entity-name="mealName"
+            v-bind="{ menu }"
+            @action="action"
+          ></context-menu>
+        </v-expand-transition>
       </v-list-item-action>
     </v-list-item>
     <v-divider></v-divider>
