+++ apps/survey/src/components/prompts/layouts/base-layout.vue
@@ -52,7 +52,7 @@
       grow
       :value="navTab"
     >
-      <review-list :meals="meals" @action="action"></review-list>
+      <meal-list-mobile :meals="meals" @action="action"></meal-list-mobile>
       <template v-if="mobileActions.length">
         <template v-for="(item, idx) in mobileActions">
           <v-btn
@@ -108,6 +108,7 @@ import type { PropType } from 'vue';
 import { defineComponent } from 'vue';
 
 import type { EncodedFood, MissingFood } from '@intake24/common/types';
+import MealListMobile from '@intake24/survey/components/layouts/meal-list/list-mobile.vue';
 import ReviewList from '@intake24/survey/components/layouts/meal-list/review-list.vue';
 
 import layoutMixin from './layout-mixin';
@@ -115,7 +116,7 @@ import layoutMixin from './layout-mixin';
 export default defineComponent({
   name: 'BaseLayout',
 
-  components: { ReviewList },
+  components: { MealListMobile },
 
   mixins: [layoutMixin],
